requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2

langchain # core frameworks
langchain-community # extra integration
faiss-cpu # Fast vector database to store the embedded data
langchain-huggingface # connect huggingface models to perform embedding
langchain-text-splitters # to split the data into smaller chunks
sentence-transformers # pre-trained models to convert chunks into vectors
langchain-core # to handle documents, chain of data etc